% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_axe.R
\name{axe_eight}
\alias{axe_eight}
\alias{axe_radon_full}
\alias{axe_radon_simul}
\alias{axe_lol}
\alias{axe_slc}
\alias{axe_air}
\title{Functions for AXE estimates}
\usage{
axe_eight(info = eight, posteriors = eight$posteriors)

axe_radon_full(info = radon_1, posteriors = radon_1$posteriors)

axe_radon_simul(
  info = radon_2,
  posteriors = radon_2$posteriors,
  var = c("axe", "iis", "map"),
  n_cores = 1,
  seed = 9871
)

axe_lol(info = lol, posteriors = lol$posteriors)

axe_slc(info = slc$data, posteriors = slc$posteriors)

axe_air(info = air$data, posteriors = air$posteriors)
}
\arguments{
\item{info}{Data to use. Varies for each, default is object generated by `prep_*()`}

\item{posteriors}{Dataframe or list with posterior mean information. Default is
object generated by `pfit_*()`.}
}
\value{
Typically a list, e.g. `eight$axe_yhats` or
    `radon_2$axe_yhats`
}
\description{
`axe_*()` A list with AXE estimates and time tkaen.
}
\section{Functions}{
\itemize{
\item \code{axe_radon_full}: Function for Radon example.

\item \code{axe_radon_simul}: Function for Radon subsets example.

\item \code{axe_lol}: Function for Esports player example.

\item \code{axe_slc}: Function for Scottish Lung Cancer example.

\item \code{axe_air}: Function for Scottish Respiratory Disease example.
}}

