% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AXEexamples-package.R, R/AXEexamples.R
\docType{package}
\name{AXEexamples-package}
\alias{AXEexamples-package}
\alias{AXEexamples}
\title{The 'AXEexamples' package.}
\description{
This R package contains data and source code to re-create the examples
in the paper "Approximate Cross-validated Mean Estimates for Bayesian Hierarchical
   Regression Models" (Zhang, Bao, Daniels 2020)

A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org

Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
