% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{df_compare_methods}
\alias{df_compare_methods}
\title{Create a dataframe which compares methods.}
\usage{
df_compare_methods(df, type = c("yhat", "elpd"))
}
\description{
Creates dataframe with LRRs, as defined in paper.
}
