% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcv.R
\name{mcv_eight}
\alias{mcv_eight}
\alias{mcv_radon_full}
\alias{mcv_radon_simul}
\alias{mcv_lol}
\alias{mcv_slc}
\alias{mcv_air}
\title{Manual cross-validation for AXE paper examples}
\usage{
mcv_eight(
  info = eight,
  n_cores = 7,
  data_scale = seq(0.1, 4, by = 0.1),
  seed = 238972,
  ...
)

mcv_radon_full(
  info = radon_1,
  models = radon_1$models_mcv,
  seed = 32897,
  n_cores = 7,
  ...
)

mcv_radon_simul(info = radon_2, seed = 9871, n_cores = 5, ...)

mcv_lol(info = lol, seed = 23982, n_cores = 7, ...)

mcv_slc(info = slc, seed = 29873, n_cores = 3, ...)

mcv_air(info = air, seed = 69872, n_cores = 3, ...)
}
\arguments{
\item{info}{All. List of data to use for CV folds; assumes list is created by
`prep_*()`.}

\item{n_cores}{All. Number of cores to use, if parallelizing. Default value varies.}

\item{seed}{Random seed generator passed to model sampling method.}

\item{...}{Additional arguments passed to modeling method.}
}
\value{
Dataframe or list of dataframes with ground-truth MCV values.

 @export
}
\description{
`mcv_*()` are high-level functions which obtain posterior predictive samples Y_j | Y_-j
    for each cross-validation loop
}
\section{Functions}{
\itemize{
\item \code{mcv_radon_full}: @export

\item \code{mcv_radon_simul}: @export

\item \code{mcv_lol}: @export

\item \code{mcv_slc}: @export

\item \code{mcv_air}: @export
}}

