% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{method_colors}
\alias{method_colors}
\title{Colors for each method in plots}
\usage{
method_colors(
  type = c("elpds", "yhats"),
  methods = c("AXE", "AXE-IIS", "AXE-MAP", "GHOST", "iIS-C", "iIS-A", "DIC-C", "DIC-A",
    "WAIC-C", "WAIC-A")
)
}
\description{
Returns named vector of colors for each method, for use in `ggplot2::scale_color_manual()`.`
}
