% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_posteriors.R
\name{pfit_eight}
\alias{pfit_eight}
\alias{pfit_lol}
\alias{pfit_slc}
\alias{pfit_air}
\title{Obtain LCO approximations for AXE paper examples}
\usage{
pfit_eight(info = eight, seed = 238972, nodes = 6, ...)

pfit_lol(info = lol, seed = 18973, ...)

pfit_slc(info = slc$data, seed = 29873, ...)

pfit_air(info = air$data, seed = 69872, ...)
}
\description{
`pfit_*()` fits models for the full data of each example in the
    AXE paper and obtains LCO approximations based on posterior samples.
}
\details{
@export
}
\section{Functions}{
\itemize{
\item \code{pfit_lol}: @export

\item \code{pfit_slc}: @export

\item \code{pfit_air}: @export
}}

