% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{plot_compare_methods}
\alias{plot_compare_methods}
\title{Plot comparison of methods (LRR)}
\usage{
plot_compare_methods(df, type = c("yhat", "elpd"))
}
\arguments{
\item{df}{Output from `df_compare_methods`}
}
\description{
Assumes use of `df_compare_methods`
}
