% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_eight}
\alias{prep_eight}
\alias{prep_radon_full}
\alias{prep_radon_simul}
\alias{prep_lol}
\alias{prep_slc}
\alias{prep_air}
\title{Data used in paper}
\usage{
prep_eight(data_scale = seq(0.1, 4, by = 0.1))

prep_radon_full()

prep_radon_simul(
  c_county = "OLMSTED",
  source = radon_1,
  data_perc = c(0.3, 0.4, 0.5, 0.6, 0.7),
  n_clusters = c(3, 4, 6, 9, 12),
  max_combn = 60,
  max_times = 1,
  seed = 9871
)

prep_lol()

prep_slc()

prep_air()
}
\arguments{
\item{data_scale}{Eight schools. Value of alpha (see paper)}

\item{c_county}{Radon subsets. Name of test county}

\item{source}{Radon subsets. List with dataframe to use (default is `radon_1`).}

\item{data_perc}{Radon subsets. Test data percentage.}

\item{n_clusters}{Radon subsets. Number of counties in data subset.}

\item{max_combn}{Radon subsets. Max number of subsets for each `data_perc` and `n_clusters.`}

\item{max_times}{Radon subsets. Number of times to iterate.}

\item{seed}{Radon subsets. Seed for subset generation.}
}
\description{
`prep_*()` returns a list with objects used to fit posteriors and obtain LCO
     approximations. See README.
}
\section{Functions}{
\itemize{
\item \code{prep_radon_full}: Function to get data for Radon example.

\item \code{prep_radon_simul}: Function to get data for Radon subsets example.

\item \code{prep_lol}: Function to get data for Esports example.

\item \code{prep_slc}: Function to get data for Scottish Lung Cancer (SLC) example.

\item \code{prep_air}: Function to get data for Scottish Respiratory Disease example.
}}

