% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{results_eight}
\alias{results_eight}
\title{Dataframe collating LCO, MCV, and AXE results}
\usage{
results_eight(
  info = eight,
  mcv_vals = eight$cv_yhats,
  axe_vals = eight$axe_yhats$axe,
  posteriors = eight$posteriors
)
}
\value{
data.frame
}
\description{
`results_*()` combines the output from `mcv_*`, `axe_*`, and `pfit_*`
    into one dataframe.
}
