% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{results_eight}
\alias{results_eight}
\alias{results_radon_full}
\alias{results_radon_simul}
\alias{results_lol}
\alias{results_slc}
\alias{results_air}
\title{Dataframe collating LCO, MCV, and AXE results}
\usage{
results_eight(
  info = eight,
  mcv_vals = eight$cv_yhats,
  axe_vals = eight$axe_yhats$axe,
  posteriors = eight$posteriors
)

results_radon_full(
  info = radon_1,
  posteriors = radon_1$posteriors,
  mcv_vals = radon_1$cv_yhats,
  axe_vals = radon_1$axe_yhats
)

results_radon_simul(
  info = radon_2,
  mcv_vals = radon_2$cv_yhats,
  axe_vals = radon_2$axe_yhats,
  posteriors = radon_2$posteriors
)

results_lol(
  info = lol,
  posteriors = lol$posteriors,
  axe = lol$axe_yhats,
  mcv_vals = lol$cv_yhats
)

results_slc(
  info = slc,
  posteriors = slc$posteriors,
  axe = slc$axe_yhats,
  mcv_vals = slc$cv_yhats
)

results_air(
  info = air,
  posteriors = air$posteriors,
  axe = air$axe_yhats,
  mcv_vals = air$cv_yhats
)
}
\value{
Dataframe

 @export
}
\description{
`results_*()` combines the output from `mcv_*`, `axe_*`, and `pfit_*`
    into one dataframe.
}
\section{Functions}{
\itemize{
\item \code{results_radon_full}: @export

\item \code{results_radon_simul}: @export

\item \code{results_lol}: @export

\item \code{results_slc}: @export

\item \code{results_air}: @export
}}

